Make the buildpack work with official Heroku multiple buildpacks
To allow the buildpack to be used with other buildpacks the way Heroku recommends, see https://devcenter.heroku.com/articles/using-multiple-buildpacks-for-an-app, the compile step now creates a `.ssh` directory in the `BUILD_DIR` and only sym-links to it for the duration of the buildpack compilation.

Furthermore, this version includes a way to customise the hosts that should be set up for `ssh` by using an environment variable. By default it's connecting to github.com as before.